Fix text disappearing after space (bug #6617)
What happened here is that when fixing bug #5734 we cahnged the
glyph image cache lookup to never return a NULL pointer, but to
instead return a 0x0 image surface. We didn't also update code
that was looking for a NULL pointer here, and as such we were
calling XRenderAddGlyphs with a 0x0 image which apparently
tickles a bug in some X servers.