[ft] Return CAIRO_STATUS_FILE_NOT_FOUND if font file can't be opened
A common source of error when people are setting up pango on a new
device is when font files don't have the right permissions and cannot
be opened.  Cairo was returning out-of-memory before, making the
Pango error message useless.

With this change, cairo will return file-not-found, and pango prints
that out.  It's still not ideal; a ENOACCESS equivalent would have
been better.

[update: Fixed a couple goofs]