Fix out of bound access in struct pattern->type
The pattern structure is hardcoded to hold 7 elements, yet the
pattern_names array in cairo-surface-observer.c is initialized with 8
strings.  This causes a crash in print_array at line 1587 when it tries
to access the 8th member.

Hence changed the 'type' array from type[7] to type[8] to avoid out of
bound access.

Fixes: https://bugs.freedesktop.org/show_bug.cgi?id=91266
Signed-off-by: Ashim <ashim.shah@samsung.com>
Reviewed-by: Bryce Harrington <bryce@osg.samsung.com>